Refactor import statements in patchifier.py and transformer.py; move append_dims function to utils.py for better organization and accessibility.
